backend/requirements.txt
Django==2.0.8
django-rest-framework==0.1.0
djangorestframework==3.8.2
Pillow==5.2.0
pytz==2018.5
